Move `Semantic` and `SymbolKind` to AST.
