+++ include/Index/SymbolIndex.h
@@ -2,8 +2,8 @@
 
 #include "Shared.h"
 #include "ArrayView.h"
+#include "AST/SymbolKind.h"
 #include "Basic/SourceCode.h"
-#include "Basic/SymbolKind.h"
 #include "Basic/RelationKind.h"
 #include "Support/JSON.h"
