+++ src/AST/SymbolKind.cpp
@@ -1,4 +1,4 @@
-#include "Basic/SymbolKind.h"
+#include "AST/SymbolKind.h"
 #include "Compiler/Compilation.h"
 
 namespace clice {
