+++ src/Feature/CodeCompletion.cpp
@@ -1,6 +1,8 @@
-#include <Basic/SourceConverter.h>
-#include <Compiler/Semantic.h>
-#include <Feature/CodeCompletion.h>
+#include "AST/SymbolKind.h"
+#include "Basic/SourceConverter.h"
+#include "Compiler/Compilation.h"
+#include "Feature/CodeCompletion.h"
+#include "clang/Sema/CodeCompleteConsumer.h"
 
 namespace clice::feature {
